Fixed calculation of ttl for event overview stage

+++ java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/shared/dispatch/ResultWithTTL.java
@@ -5,14 +5,14 @@ public class ResultWithTTL<T extends DTO> implements Result {
     /**
      * Time To Live in millisecs.
      */
-    private int ttl;
+    private long ttl;
 
     private T dto;
     
     public ResultWithTTL() {
     }
 
-    public ResultWithTTL(int ttl, T dto) {
+    public ResultWithTTL(long ttl, T dto) {
         super();
         this.ttl = ttl;
         this.dto = dto;
@@ -22,7 +22,7 @@ public class ResultWithTTL<T extends DTO> implements Result {
         return dto;
     }
 
-    public int getTtl() {
+    public long getTtl() {
         return ttl;
     }
 }
